Moved code to a new DialogueManager
