+++ The HELIOS System/Assets/Scenes/GameScene1.unity
@@ -119,6 +119,54 @@ NavMeshSettings:
     debug:
       m_Flags: 0
   m_NavMeshData: {fileID: 0}
+--- !u!1 &17396879
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 17396881}
+  - component: {fileID: 17396880}
+  m_Layer: 0
+  m_Name: DialogueManager
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &17396880
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 17396879}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: b52821d37c5d58b4693d6c3f6fcce265, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  dialogueText: {fileID: 1495581355}
+  nameText: {fileID: 112297510}
+  dialoguePanel: {fileID: 1761780117}
+  NPC: {fileID: 122572381}
+--- !u!4 &17396881
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 17396879}
+  serializedVersion: 2
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0.5707922, y: -0.28753078, z: -0.06826258}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &53596871
 GameObject:
   m_ObjectHideFlags: 0
@@ -148,10 +196,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: f043e3268329c254ba943cc6db9fb402, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  dialogueText: {fileID: 1495581355}
-  nameText: {fileID: 112297510}
-  dialoguePanel: {fileID: 1761780117}
-  NPC: {fileID: 122572381}
   organismsGoalsSerialized: []
   energyUI: {fileID: 0}
   energy: 0
@@ -991,3 +1035,4 @@ SceneRoots:
   - {fileID: 1923293001}
   - {fileID: 518381177}
   - {fileID: 122572382}
+  - {fileID: 17396881}
