+++ The HELIOS System/Assets/Scripts/DialogueManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: b52821d37c5d58b4693d6c3f6fcce265
