+++ The HELIOS System/Assets/Scripts/GameManager.cs
@@ -23,110 +23,6 @@ public class GameManager : MonoBehaviour
     public static GameManager Instance { get; private set; }
 
 
-    [Header("Dialog Management")]
-    [SerializeField] TextMeshProUGUI dialogueText;
-    [SerializeField] TextMeshProUGUI nameText;
-    [SerializeField] GameObject dialoguePanel;
-    [SerializeField] GameObject NPC;
-    public static event Action OnDialogueStarted;
-    public static event Action OnDialogueEnded;
-    private bool inDialog;
-    bool skipLineTriggered;
-
-    public void Interact()
-    {
-        Debug.Log("Interact");
-        if (inDialog)
-        {
-            Debug.Log("Skipping Line");
-            SkipLine();
-        }
-
-        else
-        {
-            inDialog = true;
-            NPCDialogueScript dialogueScript = NPC.GetComponent<NPCDialogueScript>();
-            StartDialogue(dialogueScript.dialogueAsset.dialogue, 0, dialogueScript.npcName);
-        }
-    }
-
-    public void StartDialogue(string[] dialogue, int startPosition, string name)
-    {
-        nameText.text = name + "...";
-        dialoguePanel.SetActive(true);
-        StopAllCoroutines();
-        StartCoroutine(RunDialogue(dialogue, startPosition));
-    }
-
-    IEnumerator RunDialogue(string[] dialogue, int startPosition)
-    {
-        skipLineTriggered = false;
-        OnDialogueStarted?.Invoke();
-
-        for (int i = startPosition; i < dialogue.Length; i++)
-        {
-            //dialogueText.text = dialogue[i];
-            dialogueText.text = null;
-            StartCoroutine(TypeTextUncapped(dialogue[i]));
-
-            while (skipLineTriggered == false)
-            {
-                // Wait for the current line to be skipped
-                yield return null;
-            }
-            skipLineTriggered = false;
-        }
-
-        OnDialogueEnded?.Invoke();
-        dialoguePanel.SetActive(false);
-        inDialog = false;
-    }
-
-    public void SkipLine()
-    {
-        skipLineTriggered = true;
-    }
-
-    public void ShowDialogue(string dialogue, string name)
-    {
-        nameText.text = name + "...";
-        StartCoroutine(TypeTextUncapped(dialogue));
-        dialoguePanel.SetActive(true);
-    }
-
-    public void EndDialogue()
-    {
-        nameText.text = null;
-        dialogueText.text = null;
-        dialoguePanel.SetActive(false);
-    }
-
-    float charactersPerSecond = 90;
-
-    IEnumerator TypeTextUncapped(string line)
-    {
-        float timer = 0;
-        float interval = 1 / charactersPerSecond;
-        string textBuffer = null;
-        char[] chars = line.ToCharArray();
-        int i = 0;
-
-        while (i < chars.Length)
-        {
-            if (timer < Time.deltaTime)
-            {
-                textBuffer += chars[i];
-                dialogueText.text = textBuffer;
-                timer += interval;
-                i++;
-            }
-            else
-            {
-                timer -= Time.deltaTime;
-                yield return null;
-            }
-        }
-    }
 
     [Header("Level Management System")]
 
@@ -180,7 +76,6 @@ public class GameManager : MonoBehaviour
     // Start is called once before the first execution of Update after the MonoBehaviour is created
     void Start()
     {
-        inDialog = false;
         foreach (var kvp in OrganismsSerialized)
         {
             organisms[kvp.Name] = kvp.Instance;
@@ -203,10 +98,6 @@ public class GameManager : MonoBehaviour
     void Update()
     {
         // check for player interaction with dialog
-        if (Input.GetKeyDown(KeyCode.E))
-        {
-            Interact();
-        }
     }
 
     #region Animal Management
